debug allocator is just a simple proxy in the release mode; some x64 compile time warnings fixed

+++ src/core/path.cpp
@@ -29,7 +29,7 @@ namespace Lumix
 	{
 		size_t len = path.length();
 		ASSERT(len < LUMIX_MAX_PATH);
-		PathUtils::normalize(path.c_str(), m_path, len + 1);
+		PathUtils::normalize(path.c_str(), m_path, (uint32_t)(len + 1));
 		m_id = crc32(m_path);
 	}
